feat(ui): consolidate Cluster Map + Node Status into Cluster section with tabs
Replaces three separate dashboard sections (Nodes table, Cluster Map, Node Status)
with one unified "Cluster" section containing Table / Map / Charts tabs. Active
tab persists via localStorage. Chart.js continues to lazy-load on first activation
of the Charts tab.

Adds an `embedded` prop to NodeSummaryTable, ClusterMap, and NodeStatusSection
so the parent can suppress the inner section header (icon badge, title, subtitle,
chevron) while preserving each view's functional toolbar controls (powered-off
toggle, view modes, show-predicted, grid columns, chart period).

Strips the layout preview scaffolding (LayoutPreviewPicker, NodeDeepDive, the
preview/ subdirectory with B/C/D variants) — variant A is now production.
